freeze_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
curl_cffi==0.13.0
filelock==3.19.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
hf-xet==1.1.9
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
multidict==6.6.4
multitasking==0.0.12
narwhals==2.2.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.49.0
tenacity==9.1.2
tokenizers==0.21.4
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
